src/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
databases
alembic
python-multipart
pydantic
python-dotenv
